pythonProject23/requirements.txt
Django==4.2.7
django-crispy-forms==1.13.0
djangorestframework==3.12.4
